% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.nb.point.R
\name{add.nb.point}
\alias{add.nb.point}
\title{add.nb.point will add a new column with the number of ping from the same sub trajectory contains in a circle of r radius
and in a temporal windows of temp_windows position before and temp_windows position after}
\usage{
add.nb.point(traj, r = 2000, temp_windows = 10)
}
\arguments{
\item{traj}{subset of trajectory model reduce to one traject}

\item{r}{distance to use to calculate the number of close Gps position to each track position}
}
\description{
add.nb.point will add a new column with the number of ping from the same sub trajectory contains in a circle of r radius
and in a temporal windows of temp_windows position before and temp_windows position after
}
\examples{
library(parallel)
}
