% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.extent.R
\name{create.extent}
\alias{create.extent}
\title{Tool to draw more complex extent using leaflet}
\usage{
create.extent(gps.extent)
}
\arguments{
\item{traj}{A subset of global trajectory dataset identified by one no_trajet}
}
\description{
Tool to draw more complex extent using leaflet
}
\examples{
#emprise<-create.extent(st_convex_hull(st_union(gps.all)))
}
