% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.predict.R
\name{glm.predict}
\alias{glm.predict}
\title{Prediction on fishing activity based on a given model}
\usage{
glm.predict(traj, gear.glm, seuil)
}
\arguments{
\item{traj}{A subset of global trajectory dataset identified by one no_trajet}

\item{gear.glm}{The glm model used for the prediction}

\item{seuil}{Threshold used to predict fishing activity}
}
\description{
Prediction on fishing activity based on a given model
}
\examples{
data(R2)
engin_encours<-"FMCy"
gear.glm<-model.traj.glm(filter(R2,code_engin==engin_encours),observation='activity')
R2.pred<-glm.predict(filter(R2,code_engin==engin_encours),gear.glm,seuil=0.5)
ggplot(R2.pred)+geom_sf(size=0.1,col='red')+ggtitle("GPS ping without prediction")
ggplot(R2.pred)+geom_sf(size=0.1,aes(col=activity_plus))+ggtitle("GPS ping with GLM prediction")

}
