% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPX.load.R
\name{GPX.load}
\alias{GPX.load}
\title{Load one GPX file or All gpx files of a give directory}
\usage{
GPX.load(file = NULL, gpx_rep = NULL)
}
\arguments{
\item{gpx_rep}{Name of the directory that contains all gps files that should be loaded}

\item{files}{Name of a gpx file to be loaded}
}
\description{
Load one GPX file or All gpx files of a give directory
}
\examples{
#GPX.load(gpx_rep='data/gpx')
}
