% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltraj2sf.R
\name{ltraj2sf}
\alias{ltraj2sf}
\title{ltraj2sf}
\usage{
ltraj2sf(ltrajobj)
}
\arguments{
\item{ltrajobj}{ltraj object produced when redistribute trajectory to have constant duration beetween pings}
}
\description{
ltraj2sf
}
\examples{

}
