% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.traj.RF.R
\name{model.traj.RF}
\alias{model.traj.RF}
\title{Tool to draw more complex extent using}
\usage{
model.traj.RF(traj, observation = NULL, form = NULL)
}
\arguments{
\item{traj}{A subset of global trajectory dataset identified by one no_trajet}

\item{obervation}{The name of the column used to store observed activity}

\item{formula}{Default formula is observation~dist but users can specified his on model}
}
\description{
Tool to draw more complex extent using
}
\examples{
#gear.glm<-model.traj.RF(filter(R2,code_engin==engin_encours),observation='activity','dist')
}
