% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_isolated.R
\name{remove_isolated}
\alias{remove_isolated}
\title{Tool to draw more complex extent using}
\usage{
remove_isolated(vect_fish, state, replacement, lim = 10)
}
\arguments{
\item{vect_fish}{vector of fishing event}

\item{state}{event to focus on}

\item{replacement}{threshold, number of point to be considered as isolated}
}
\description{
Tool to draw more complex extent using
}
