% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.predict.R
\name{RF.predict}
\alias{RF.predict}
\title{Prediction on fishing activity based on a given model}
\usage{
RF.predict(traj, mod.RF)
}
\arguments{
\item{traj}{A subset of global trajectory dataset identified by one no_trajet}

\item{mod.RF}{The Random forest model calculate by ranger function}
}
\description{
Prediction on fishing activity based on a given model
}
\examples{
#glm.predict(filter(R2,code_engin==engin_encours),gear.glm2,seuil=0.5)
}
