% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silicotraj.R
\name{silicoTraj}
\alias{silicoTraj}
\title{Propose a user interface to identified portion of the tracks that could be active one.}
\usage{
silicoTraj(traj, mode = "map")
}
\arguments{
\item{traj}{A subset of global trajectory dataset identified by one no_trajet}
}
\description{
Propose a user interface to identified portion of the tracks that could be active one.
}
\examples{
\dontrun{
data(R2)

R2$activity_plus<-'UK'
new_trajet29<-silicoTraj(dplyr::filter(R2,no_trajet==29),mode='speed')

new_trajet29<-silicoTraj(dplyr::filter(R2,no_trajet==29),mode='map')
}
}
