% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_replace.R
\name{track_replace}
\alias{track_replace}
\title{Replace a given track id by a new one. Espacially used when adding In silico observation to a given set of trajectory.}
\usage{
track_replace(trajInit, new_track)
}
\arguments{
\item{trajInit}{A big sf objection with all the track id used}

\item{new_track}{A subset of trajInit modified and to insert into the dataset instead of the old one}
}
\description{
Replace a given track id by a new one. Espacially used when adding In silico observation to a given set of trajectory.
}
\examples{

}
