% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.add.nb.point.R
\name{all.add.nb.point}
\alias{all.add.nb.point}
\title{all.add.nb.point will proceed to add.nb.point to all subtrajectory of a global traj dataset}
\usage{
\method{all}{add.nb.point}(traj, r = 2000, temp_windows = 10)
}
\arguments{
\item{traj}{dataset of trajectories}

\item{r}{distance to use to calculate the number of close Gps position to each track position}
}
\description{
all.add.nb.point will proceed to add.nb.point to all subtrajectory of a global traj dataset
}
\examples{
data(R2)
R2test_retour<-all.add.nb.point <-function (R2,r=2000)
}
