% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.extent.R
\name{create.extent}
\alias{create.extent}
\title{Tool to draw more complex extent using leaflet}
\usage{
create.extent(gps.extent)
}
\arguments{
\item{traj}{A subset of global trajectory dataset identified by one no_trajet}
}
\description{
Tool to draw more complex extent using leaflet
}
\examples{
\dontrun{
library(GPSMonitoring)
library(dplyr)
library(sf)
data(GPSdataset)

GPSdataset \%>\% mutate(filename=paste(code_village,code_engin,code_pecheur,'.gpx',sep='_')) \%>\%
arrange (filename) \%>\% dplyr::distinct(code_village,code_engin,code_pecheur,filename) \%>\%
 dplyr::mutate(track_fid=row_number(),track_seg_id=track_fid) \%>\%
 inner_join(GPSdataset) \%>\%
 group_by(filename) \%>\% arrange (filename,date_heure) \%>\% dplyr::mutate(track_seg_point_id = row_number()) \%>\%
 dplyr::rename(time=date_heure) \%>\%
 st_as_sf(coords = c("longitude", "latitude"), crs = 4326,remove=FALSE)->gps.all
emprise<-create.extent(st_convex_hull(st_union(gps.all)))
}
}
