% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.predict.R
\name{glm.predict}
\alias{glm.predict}
\title{Prediction on fishing activity based on a given model}
\usage{
glm.predict(traj, gear.glm, seuil)
}
\arguments{
\item{traj}{A subset of global trajectory dataset identified by one no_trajet}

\item{gear.glm}{The glm model used for the prediction}

\item{seuil}{Threshold used to predict fishing activity}
}
\description{
Prediction on fishing activity based on a given model
}
\examples{
#glm.predict(filter(R2,code_engin==engin_encours),gear.glm,seuil=0.5)
}
