% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.add_traj_number.R
\name{GPS.add_traj_number}
\alias{GPS.add_traj_number}
\title{GPS.add_traj_number, the function will split set of ddata collected from a file in several independant trajectory.
A set of ping is consider as a new trajectory if duration between two ping is higher that limit parameter}
\usage{
GPS.add_traj_number(GPS.data, limit)
}
\arguments{
\item{GPS.data}{sf object set up for GPX.load function}

\item{limit}{definition of the duration between two position to consider a new traject}
}
\description{
GPS.add_traj_number, the function will split set of ddata collected from a file in several independant trajectory.
A set of ping is consider as a new trajectory if duration between two ping is higher that limit parameter
}
\examples{

}
