% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.curation.R
\name{GPS.curation}
\alias{GPS.curation}
\title{Generic function to cure gps data}
\usage{
GPS.curation(GPS.data, extent = NULL, exclude = NULL)
}
\arguments{
\item{GPS.data}{sf object set up for GPX.load function}

\item{extent}{sf object for area of interest definition}

\item{exclude}{sf object for exclusion of data (ie port by examples)}
}
\description{
Generic function to cure gps data
}
