% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.traj.glm.R
\name{model.traj.glm}
\alias{model.traj.glm}
\title{GLM model on the data}
\usage{
model.traj.glm(traj, observation = NULL, form = NULL)
}
\arguments{
\item{traj}{A subset of global trajectory dataset identified by one no_trajet}

\item{obervation}{The name of the column used to store observed activity}

\item{formula}{Default formula is observation~dist but users can specified his on model}
}
\description{
GLM model on the data
}
