% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.ind.R
\name{quality.ind}
\alias{quality.ind}
\title{Calculate and print sensitivity, accuracy and}
\usage{
quality.ind(proj.traj, col.activity = NULL, col.predict = NULL)
}
\arguments{
\item{col.activity}{Column name for observed fishing event}

\item{col.predict}{Column name for predict values}

\item{proj.tarj}{Set of trajectory with predict.glm values}
}
\description{
Calculate and print sensitivity, accuracy and
}
\examples{
#GPX.load(gpx_rep='data/gpx')
}
