% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redis.traj.R
\name{Redis.traj}
\alias{Redis.traj}
\title{Redis.traj function will redistribute ping along the trajectory with a constant duration between ping with respect of speed and time of initial data.}
\usage{
Redis.traj(GPS.data, step_dt, silent = NULL)
}
\arguments{
\item{GPS.data}{sf object loading GPX.load function or similar one}

\item{step}{constant between ping}
}
\description{
Redis.traj function will redistribute ping along the trajectory with a constant duration between ping with respect of speed and time of initial data.
}
\examples{

}
