% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.predict.R
\name{RF.predict}
\alias{RF.predict}
\title{Prediction on fishing activity based on a given model}
\usage{
RF.predict(traj, mod.RF)
}
\arguments{
\item{traj}{A subset of global trajectory dataset identified by one no_trajet}

\item{mod.RF}{The Random forest model calculate by ranger function}
}
\description{
Prediction on fishing activity based on a given model
}
\examples{
data(R2)
library(ranger)
engin_encours<-"FMCy"
gear.RF<-model.traj.RF(filter(R2,code_engin==engin_encours),observation='activity','dist')
R2.pred<-RF.predict(filter(R2,code_engin==engin_encours),gear.RF)
ggplot(R2.pred)+geom_sf(size=0.1,col='red')+ggtitle("GPS ping without prediction")
ggplot(R2.pred)+geom_sf(size=0.1,aes(col=activity_plus))+ggtitle("GPS ping with RF prediction")
}
